refactor(storage): move record persistence behind storage
